Minor fixes

+++ Install-OhMyPosh.ps1
@@ -3,7 +3,11 @@ Param()
 
 Install-WinGetApp -App "JanDeDobbeleer.OhMyPosh" -ErrorAction Stop
 
-$themes = Get-ChildItem -Path "$PSScriptRoot\oh-my-posh\themes" -Recurse "*.json"
-foreach ($theme in $themes) {
-    Copy-Item $theme.FullName -Destination $env:POSH_THEMES_PATH -Force
+Write-Host "Loading oh-my-posh themes..." 
+$env:POSH_THEMES_PATH = [System.Environment]::GetEnvironmentVariable("POSH_THEMES_PATH", "User") 
+if (Test-Path $env:POSH_THEMES_PATH) {
+    $themes = Get-ChildItem -Path "$PSScriptRoot\oh-my-posh\themes" -Recurse "*.json"
+    foreach ($theme in $themes) {
+        Copy-Item $theme.FullName -Destination $env:POSH_THEMES_PATH -Force
+    }
 }
